app/requirements.txt
# ===== 核心框架 =====
langgraph>=0.2.0
langchain>=0.3.0
langchain-community>=0.3.0
chainlit>=1.3.0

# ===== LLM 接口 =====
dashscope>=1.20.0            # 阿里云 Qwen SDK
openai>=1.0.0                # OpenAI 兼容接口 (备用)

# ===== RAGFlow 检索 (通过 HTTP API) =====
# httpx 已在下方工具库中声明

# ===== 评测 (ragas 通过 fork 安装，此处不声明) =====
# ragas 已在 Dockerfile 中从 vendor/ragas 安装

# ===== 工具库 =====
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.27.0
